Разбивка кода и единый стракт настроек
